scripts: aerospace ghost window fix
This is a simple script that looks for any "ghost" windows in aerospace
and then closes them out without closing out the program.

Created due to "ghost" windows appearing in aerospace while running the
MacOS 26 developer beta. Only occurs when window is closed through a
method that is not aerospace.