nvim: update plugins reliant on themes to account for system appearance
Using auto-dark-mode.lua, we can set a global variable to use in other
plugins that will determine the systems current light/dark theme.
An autocmd is used to then check for when auto-dark-mode updates that
variable as it's preforming a constant check.

This isn't perfect as this only checks the systems theme. If you are not
using a terminal that also changes it's theme depending on the system
appearance it will make nvims colors unusable. Should look into checking
the terminal theme instead of the system theme, but this works for me
since I use Ghostty or iTerm2 and have it set to change themes depending on the
system appearance.